List details of **all** MCA Labeling labels
We previously listed the MCA Tag Name and MCA Tag Symbol for the labels
that we defined. We now also list those that are defined elsewhere (IMF,
DCI) because we need to apply prefixes to the MCA Tag Symbols for the
externally-defined labels.

This commit also applies those same prefixes to MCA Tag Symbols of the
lables that we are defining, so the MCA Tag Symbols are consistent.